rename executable
