+++ libs/cvr-fixture-generator/src/cli/generate/index.ts
