+++ libs/cvr-fixture-generator/src/cli/generate/main.test.ts
