+++ libs/cvr-fixture-generator/src/cli/generate/main.ts
@@ -28,7 +28,7 @@ import {
  * Script to generate a cast vote record file for a given election.
  * Run from the command-line with:
  *
- * ./bin/generate-sample-cvr-file -h
+ * ./bin/generate -h
  *
  * To see more information and all possible arguments.
  */
