+++ libs/cvr-fixture-generator/src/index.ts
@@ -1,2 +1,2 @@
 export * from './generate_cvrs';
-export * as GenerateSampleCvrFileCli from './cli/generate-sample-cvr-file';
+export * as GenerateCli from './cli/generate';
